Set up pre-test deletion and env-specific post-test deletion for queries

+++ src/ecomm-api-svc/cypress/integration/queries/productAttributes.spec.ts
@@ -146,8 +146,19 @@ describe('Query: productAttributes', () => {
             });
         };
 
+        var deleteItemsAfter = undefined as boolean | undefined;
+        before(() => {
+            deleteItemsAfter = Cypress.env("deleteItemsAfter");
+            cy.deleteCypressItems(queryName, deleteName).then(() => {
+                cy.deleteCypressItems(productQuery, productDelete, "productInfo", `Cypress ${queryName}`);
+            });
+        });
+
         // Ensure deletion of the items we created for the productId test
         after(() => {
+            if (!deleteItemsAfter) {
+                return;
+            }
             cy.deleteSupplementalItems(createdProducts);
             cy.deleteSupplementalItems(createdItems);
         });
